refactor: 디렉토리, 파일명 변경
