+++ src/common/template.js
